Move bin logic to a module. Fixes #29
